Refactor remover.sh to check for resource existence before removal of key vault secrets

+++ deploy/scripts/remover.sh
@@ -274,7 +274,6 @@ echo "Target subscription:                 ${ARM_SUBSCRIPTION_ID}"
 echo "Deployer State file:                 ${deployer_tfstate_key}"
 echo "Landscape State file:                ${landscape_tfstate_key}"
 
-
 export TF_VAR_subscription_id="${ARM_SUBSCRIPTION_ID}"
 # Terraform Plugins
 if checkIfCloudShell; then
@@ -485,32 +484,52 @@ if [ "$resource_group_exist" ]; then
 		allParameters=$(printf " -var-file=%s %s %s  %s " "${var_file}" "${extra_vars}" "${tfstate_parameter}" "${deployer_tfstate_key_parameter}")
 
 		moduleID="module.sap_landscape.azurerm_key_vault_secret.sid_ppk"
-		if terraform -chdir="${terraform_module_directory}" state rm "${moduleID}"; then
-			echo "Secret 'sid_ppk' removed from state"
+		if terraform -chdir="${terraform_module_directory}" state list -id="${moduleID}"; then
+			if terraform -chdir="${terraform_module_directory}" state rm "${moduleID}"; then
+				echo "Secret 'sid_ppk' removed from state"
+			fi
 		fi
+
 		moduleID="module.sap_landscape.azurerm_key_vault_secret.sid_pk"
-		if terraform -chdir="${terraform_module_directory}" state rm "${moduleID}"; then
-			echo "Secret 'sid_pk' removed from state"
+		if terraform -chdir="${terraform_module_directory}" state list -id="${moduleID}"; then
+			if terraform -chdir="${terraform_module_directory}" state rm "${moduleID}"; then
+				echo "Secret 'sid_pk' removed from state"
+			fi
 		fi
-		moduleID="module.sap_landscape.azurerm_key_vault_secret.sid_username"
-		if terraform -chdir="${terraform_module_directory}" state rm "${moduleID}"; then
-			echo "Secret 'sid_username' removed from state"
+
+		if terraform -chdir="${terraform_module_directory}" state list -id="${moduleID}"; then
+			moduleID="module.sap_landscape.azurerm_key_vault_secret.sid_username"
+			if terraform -chdir="${terraform_module_directory}" state rm "${moduleID}"; then
+				echo "Secret 'sid_username' removed from state"
+			fi
 		fi
+
 		moduleID="module.sap_landscape.azurerm_key_vault_secret.sid_password"
-		if terraform -chdir="${terraform_module_directory}" state rm "${moduleID}"; then
-			echo "Secret 'sid_password' removed from state"
+		if terraform -chdir="${terraform_module_directory}" state list -id="${moduleID}"; then
+			if terraform -chdir="${terraform_module_directory}" state rm "${moduleID}"; then
+				echo "Secret 'sid_password' removed from state"
+			fi
 		fi
+
 		moduleID="module.sap_landscape.azurerm_key_vault_secret.witness_access_key"
-		if terraform -chdir="${terraform_module_directory}" state rm "${moduleID}"; then
-			echo "Secret 'witness_access_key' removed from state"
+		if terraform -chdir="${terraform_module_directory}" state list -id="${moduleID}"; then
+			if terraform -chdir="${terraform_module_directory}" state rm "${moduleID}"; then
+				echo "Secret 'witness_access_key' removed from state"
+			fi
 		fi
+
 		moduleID="module.sap_landscape.azurerm_key_vault_secret.deployer_keyvault_user_name"
-		if terraform -chdir="${terraform_module_directory}" state rm "${moduleID}"; then
-			echo "Secret 'deployer_keyvault_user_name' removed from state"
+		if terraform -chdir="${terraform_module_directory}" state list -id="${moduleID}"; then
+			if terraform -chdir="${terraform_module_directory}" state rm "${moduleID}"; then
+				echo "Secret 'deployer_keyvault_user_name' removed from state"
+			fi
 		fi
+
 		moduleID="module.sap_landscape.azurerm_key_vault_secret.witness_name"
-		if terraform -chdir="${terraform_module_directory}" state rm "${moduleID}"; then
-			echo "Secret 'witness_name' removed from state"
+		if terraform -chdir="${terraform_module_directory}" state list -id="${moduleID}"; then
+			if terraform -chdir="${terraform_module_directory}" state rm "${moduleID}"; then
+				echo "Secret 'witness_name' removed from state"
+			fi
 		fi
 
 		if [ -n "${approve}" ]; then
